DAT-1180 reports without conditions
