+++ cmdb/security/acl/control.py
@@ -14,15 +14,21 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 """TODO: document"""
+from typing import TypeVar
+
 from cmdb.security.acl.permission import AccessControlPermission
-from cmdb.security.acl.sections import GroupACL, T
+from cmdb.security.acl.group_acl import GroupACL
 # -------------------------------------------------------------------------------------------------------------------- #
 
+T = TypeVar('T')
+
+# -------------------------------------------------------------------------------------------------------------------- #
+#                                               AccessControlList - CLASS                                              #
+# -------------------------------------------------------------------------------------------------------------------- #
 class AccessControlList:
     """
     The actual implementation of the Access Control List (ACL).
     """
-
     def __init__(self, activated: bool, groups: GroupACL = None):
         self.activated = activated
         self.groups: GroupACL = groups
